% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_colors.R
\name{nyc_palette}
\alias{nyc_palette}
\title{Color Palettes based on NYC}
\usage{
nyc_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Choices are:
\code{ParkBlossoms}, \code{SunsetPlatform}, \code{EmpireView}, \code{Sunset1},
\code{Sunset2}, \code{Skyline1}, \code{Skyline2}, \code{ShoreSkyline}, \code{Empire},
\code{Statue1}, \code{Statue2}, \code{World1}, \code{World2}, \code{Met}, \code{BrooklynBridge},
\code{DUMBO}, \code{Yankees}}

\item{n}{Number of colors desired. Palettes range from 4 - 7 colors.}

\item{type}{Either continuous or discrete.}
}
\value{
A vector of colors.
}
\description{
R package that contains color palettes based on pictures that I took around NYC.
}
\examples{
nyc_palette("Sunset2")
nyc_palette("DUMBO")

}
